refactor: extract internal list components and add GroupBy prop
