+++ lib/components/ChangelogList/ChangelogList.types.ts
@@ -1,0 +1,4 @@
+export enum GroupBy {
+  COMPONENT = 'COMPONENT',
+  NONE = 'NONE',
+}
