+++ lib/components/ChangelogList/index.ts
@@ -1,2 +1,4 @@
+export { GroupBy } from './ChangelogList.types';
+
 export { MinimalChangelogList } from './MinimalChangelogList';
 export { ChangelogList } from './ChangelogList';
